Code review updated.
