+++ src/AwesomeCMSCore/AwesomeCMSCore/React/css/Group/Index.scss
@@ -1,3 +1,3 @@
-﻿#groupTable{
-    margin:5%;
+﻿#groupTable {
+    margin: 5%;
 }
