+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Account/Repositories/AccountRepository.cs
@@ -55,9 +55,7 @@ namespace AwesomeCMSCore.Modules.Account.Repositories
                                       Email= user.Email,
                                       EmailConfirmed= user.EmailConfirmed.ToString() 
                                       ,Groups = user.Groups.ToArray().ToString() 
-                                  }).ToListAsync();
-
-            
+                                  }).ToListAsync(); 
             return userList;
         }
