+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Entities/Entities/ApplicationGroup.cs
@@ -5,23 +5,10 @@ using System.ComponentModel.DataAnnotations;
 using System.Text;
 
 namespace AwesomeCMSCore.Modules.Entities.Entities
-{
-   
-    public class ApplicationGroupRole
-    {
-        [Required]
-        public string RoleId { get; set; }
-        [Required]
-        public string GroupId { get; set; }
-        public ApplicationRole Role { get; set; }
-        public ApplicationGroup Group { get; set; } 
-    }
-
+{ 
     public class ApplicationGroup
     {
-        public ApplicationGroup() { }
-
-
+        public ApplicationGroup() { } 
         public ApplicationGroup(string name) : this()
         {
             this.Roles = new List<ApplicationGroupRole>();
@@ -34,18 +21,5 @@ namespace AwesomeCMSCore.Modules.Entities.Entities
         public string Name { get; set; }
         public ICollection<ApplicationGroupRole> Roles { get; set; }
         public ICollection<ApplicationUserGroup> Users { get; set; }
-    }
-
-
-
-    public class ApplicationUserGroup
-    {
-        [Required]
-        public string UserId { get; set; }
-        [Required]
-        public string GroupId { get; set; }
-        public User User { get; set; }
-        public ApplicationGroup Group { get; set; }
-    }
-
+    } 
 }
